Fix 1.10 entity data tracker issues

+++ src/main/java/net/earthcomputer/multiconnect/protocols/v1_10/Protocol_1_10.java
@@ -412,7 +412,7 @@ public class Protocol_1_10 extends Protocol_1_11 {
         if (clazz == HorseBaseEntity.class && data == HorseBaseEntityAccessor.getHorseFlags()) {
             DataTrackerManager.registerOldTrackedData(HorseBaseEntity.class, OLD_HORSE_FLAGS, (byte)0, (entity, val) -> {
                 // keep the bottom 3 flags, skip the 4th and shift the higher ones down
-                entity.getDataTracker().set(HorseBaseEntityAccessor.getHorseFlags(), (val & 7) | ((val & ~15) >>> 1));
+                entity.getDataTracker().set(HorseBaseEntityAccessor.getHorseFlags(), (byte) ((val & 7) | ((val & ~15) >>> 1)));
                 if (entity instanceof AbstractDonkeyEntity) {
                     ((AbstractDonkeyEntity) entity).setHasChest((val & 8) != 0);
                 }
