+++ src/main/java/net/earthcomputer/multiconnect/protocols/v1_10/mixin/HorseBaseEntityAccessor.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(HorseBaseEntity.class)
 public interface HorseBaseEntityAccessor {
     @Accessor("HORSE_FLAGS")
-    static TrackedData<Integer> getHorseFlags() {
+    static TrackedData<Byte> getHorseFlags() {
         return MixinHelper.fakeInstance();
     }
 }
